backend/requirements.txt
aniso8601==6.0.0
astroid==2.2.5
atomicwrites==1.4.1
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
Click==7.0
colorama==0.4.6
cryptography==39.0.2
defusedxml==0.7.1
ecdsa==0.13.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.4
future==0.17.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.2.2
packaging==23.1
pluggy==0.13.1
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.3.1
PyJWT==2.6.0
pylint==2.3.1
pytest==6.2.2
python-dotenv==1.0.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.16.5
six==1.12.0
social-auth-core==4.3.0
SQLAlchemy==1.4.2
toml==0.10.2
typed-ast==1.4.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==0.15.6
wrapt==1.11.1